More work on Modification and resulting other data class changes

+++ Proto/core/src/squidpony/epigon/data/mixin/Container.java
@@ -1,11 +1,12 @@
 package squidpony.epigon.data.mixin;
 
 import java.util.List;
+
 import squidpony.epigon.data.specific.Physical;
 
 public class Container {
 
-    public int capacity;
+    public double capacity;
 
     public List<Physical> contents;
 }
